requirements.txt
streamlit==1.22.0
pandas==2.0.1
requests==2.28.0
ydata-profiling==4.3.1
matplotlib==3.6.0
mysql-connector-python==8.0.33
autoviz
ipython
klib